requirements.txt
asgiref==3.7.2
Django==4.2.5
django-ckeditor==6.7.0
django-environ==0.11.2
django-js-asset==2.1.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
phonenumbers==8.13.22
Pillow==10.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
